+++ src/Query/Grammars/PostgresGrammar.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Php\Support\Laravel\Database\Query\Grammars;
 
 use Illuminate\Database\Query\Grammars\PostgresGrammar as BasePostgresGrammar;
@@ -10,7 +12,7 @@ class PostgresGrammar extends BasePostgresGrammar
     {
         $cols = array_unique(array_filter($cols));
 
-        $returns = implode(',', array_map(fn($item) => $this->wrap($item), $cols));
+        $returns = implode(',', array_map($this->wrap(...), $cols));
 
         return $returns ? " RETURNING $returns" : '';
     }