+++ src/Schema/Postgres/Grammar/GrammarIndexes.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace Php\Support\Laravel\Database\Schema\Postgres\Grammar;
 
-use Php\Support\Laravel\Database\Schema\Postgres\Blueprint;
 use Php\Support\Laravel\Database\Schema\Postgres\Builders\Indexes\Unique\UniqueBuilder;
+use Illuminate\Database\Schema\Blueprint as BaseBlueprint;
 use Php\Support\Laravel\Database\Schema\Postgres\Builders\Indexes\PartialBuilder;
 use Php\Support\Laravel\Database\Schema\Postgres\Compilers\PartialCompiler;
 use Php\Support\Laravel\Database\Schema\Postgres\Compilers\UniqueCompiler;
 
 trait GrammarIndexes
 {
-    public function compileUniquePartial(Blueprint $blueprint, UniqueBuilder $command): string|array
+    public function compileUniquePartial(BaseBlueprint $blueprint, UniqueBuilder $command): string|array
     {
         $constraints = $command->get('constraints');
 
@@ -26,7 +26,7 @@ trait GrammarIndexes
         return $this->compileUnique($blueprint, $command);
     }
 
-    public function compilePartial(Blueprint $blueprint, PartialBuilder $command): string
+    public function compilePartial(BaseBlueprint $blueprint, PartialBuilder $command): string
     {
         return PartialCompiler::compile($this, $blueprint, $command);
     }